[Bronze V] Title: A+B - 9, Time: 172 ms, Memory: 17760 KB, Score: 75 point -BaekjoonHub

+++ 백준/Bronze/15740. A＋B － 9/A＋B － 9.java
@@ -4,8 +4,8 @@ public class Main {
     
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
         System.out.print(a+b);
     }
